+++ actions/ErrorAction.php
@@ -80,7 +80,7 @@ class ErrorAction extends \yii\web\ErrorAction
             return (array) $rr;
         } else
         {
-            if (\Yii::$app->admin->requestIsAdmin)
+            if (\Yii::$app->admin->checkAccess() && \Yii::$app->admin->requestIsAdmin)
             {
                 if (\Yii::$app->user->can(CmsManager::PERMISSION_ADMIN_ACCESS))
                 {